split non working tests
